initial implementation to escape string IDs
String IDs are these strings that are later used in AQL queries, e.g. annotation names or edge names. Since AQL does limits the characters that can be included as names we should convert invalid IDs already when converting to ANNIS. The database would have no problem of storing the information but the annotation could not be queried and thus be useless.